Make SINK() preserve node validity invariant of the incoming cell
SINK() is good because it means you're declaring that a value cell is
going to be overwritten, and you don't know how long that writing
process is going to take.  So it's best if you make it noisy/bad if
someone tries to use any of those fields before that write happens.

Also by overwriting the payload bits, it makes sure you don't
incidentally/accidentally wind up carrying over any bits from the
previous value (which could be any type) to the new one.

But the "trash" status of a cell prevents it from being legal to be
seen by the GC.  As it's not known how long a write to a target cell
is going to take, the trashing state could span a GC...and the original
cell may have been intended to be GC safe.

A previous patch simply made the GC tolerate trash.  But this loses
a fairly important sanity check.

So this commit instead changes SINK() to not violate the invariant of
the cell you got in.  If it was valid for GC before, it should stay
valid for GC during the period before it's overwritten.  And if it was
invalid for GC, it should stay invalid.

Hence valid cells are turned to UNREADABLE_BLANK, while trash cells
are left as trash.

Includes fixing a test that was added in the previous commit that did
not return a LOGIC! result