Fixes related to audit of stack-volatile pointers across DS_PUSH
The policy in Ren-C is that a DS_PUSH (or DS_POP) may invalidate any
pointer into the data stack--because it is a contiguously allocated
array that might have to be moved to fulfill the request.

This was stress-tested using some code that would relocate the data
stack array on every DS_PUSH, invalidating all the old memory cells
so that any usage of them would trigger a crash.

Only a couple of violations were discovered.  This brings them into
compliance.

Also includes a patch for "printf" to indicate it does not return a
result, which fixes ls/l and ls/r