Do not let TRACE apply during running of the REPL code itself
Due to having some of the REPL running as Rebol, it becomes a bit
of a problem with TRACE, since the way trace works is a simple global
flag (two flags, really)...and when you start adding in debug spew,
the user winds up tracing into the REPL itself, since the setting of
that flag spans its code.

As a quick way to address this particular problem, the trace flags are
cached by the host before running the REPL, and restored before running
the user's actual code.