Use printf to output file and line number info in a panic
As a debug feature, panic relies on triggering a kind of crash that
gets Valgrind or address sanitizer to give detailed information about
the origin of series or values.  However, this triggered crash will
end the program for those not using a memory tool, and it was doing
so before the file or line number was output.

printf is not included in the release build, so that's not how panic
messages are delivered.  But in particular with the file and line
information, that data will only be in the debug build--which is
allowed to use printf.  So it's best to early on in the panic just
print the file and line number and flush the output buffer.

Includes a tweak to HELP so it works on functions.