Try Travis building where GIT_COMMIT is unspecified
On the platforms that build twice, this makes the first build not
specify a GIT_COMMIT, since that is a build variation which would not
otherwise be tested by Travis.  Then on the second build using the
generated executable, it does specify a GIT_COMMIT, so that the final
executable will have one.

(This should currently fail, as a recent merge merged a PR which broke
the no-commit behavior.)