Avoid calling Error() directly
Because Error is a variadic func, the compiler can't check the
number or type of the arguements. The check can only happen at runtime,
due to the nature of the error handling, it might be tested at all.

Let make-boot.r generate a wrapper function for each error code, and
Only use these wrappers in the code, so that the compiler can do some
checking. To distinguish the auto-generated wrappers and the ones
defined in c-error.c, the name of auto-generated ones ends with _Raw().

This did catch a couple of problems in error construction.