Base %make-make.r on common emitter, add more options
To keep bootstrapping Rebol on new systems easy, there is a makefile
generator, and the process of building all the generated include files
can be performed on a different machine than the one being built on.
The only requirement for the new machine is to have GNU make, which is
an old and relatively simple tool.

The generation process was based on doing search and replaces in a
template makefile string, with another part being done in an ordinary
emit fashion.  This somewhat limited the generation process, meaning
any changes had to be done in a fairly indirect and confusing way.

This changes the generation to be emit-based to do parameterization.
That means that instead of having a false makefile that gets overridden,
the precise places to make the substitution are done with code.  This
allowed the extension building code to use the same command line
interface as CMake, instead of having two different conventions.

Initially, the system was only set up to take one parameter: OS_ID.
This brings more parameterization, including some of the stuff done
manually in Travis, such as static linking.