Use union for void* vs. CFUNC* in HANDLE!
The sizes of void* and C function pointers are not required to be the
same in C on all architectures.  Typically it is the case, but it's
not generally a requirement for Rebol to work.

The HANDLE! can hold either a function pointer or a data pointer, so
this switches it to a union.  A flag also remembers which kind of
handle it actually is.

Only the FFI interface requires void* and CFUNC* to really be the
same size, due to mislabeling an argument as a void* which really
should be a function pointer.  To get past strict checking which will
really disallow you from doing a void* to function cast, this uses
a memcpy to work around the generically illegal conversion.