Fix undefined behaviors for INTEGER!/DECIMAL! bit access
For reasons that are not entirely clear, some R3-Alpha code was
accessing the bytes of decimals by way of casting it to an integer.
This is a violation of strict aliasing.

To solve the problem, this does the more obvious accessing of bytes
via a `char*`, which is also exempt from strict aliasing rules.

Also contains a correction for a dicey coercion that was used to
inject an assert on VAL_INTEGER() that the value was actually an
integer.  This was tricky because of the desire to write
`VAL_INTEGER(x) = 10` as well as to be able to use it on const
values.  Since there's no clean way to do this check in the C build,
the checks are deferred to be in the C++ build only...and a similar
check is added into VAL_DECIMAL() and VAL_INDEX().