Change some tmp-*.inc files to just tmp-*.c
Some of the generated tables are global data, and hence could only
be included once in the entire project.  To make this clearer, they
had been renamed to ".inc" files instead of ".h" some time ago.

It was difficult to decide which other core file exactly "owned" the
.inc file to be the dedicated one to include it.  This created some
frustrations, so %sys-core.h added in extern prototypes for the
tables so anyone could access them.

Doing so leads to a double-definition in the file that does the .inc.
Strict compilation complains about that, so the easiest thing to do
is just move them into .c files and have them include %sys-core.h.