Turn on the PROTECTing working on function args/locals
Due to the fact that not all the bits in a value header are copied when
Move_Value is done, it's possible to set the protection status of a
variable on the value vs. the key.  This means the keylist does not
have to be modified, and hence it doesn't have to be made unique
from any objects that were sharing it.

While the basic mechanic been implemented, the protection routines had
not had the uniqueness removed if the case was a PROTECT.  The
PROTECT/HIDE feature still uses typeset keys.