Fix CMake path for %make-headers.r
The somewhat convoluted history of the make process meant there was
some confusing code regarding where to look for the files to scan
for natives.  A simplification confused the CMake build, so that it
was looking for source files in the generated files directory.

There is at this time no need to scan the generated files for headers,
so the path modification is not necessary.  The generated files
are now in their own list in %file-base.r