BIG_ENDIAN/LITTLE_ENDIAN => ENDIAN_BIG/ENDIAN_LITTLE
This fixes broken binary conversions on Windows introduced when
getting rid of XXX_DECIMAL_BITS() routines.

The defines used by Rebol are ENDIAN_XXX and not XXX_ENDIAN.  There is
apparently a reason for this, and that reason affected the fact that
a mistake in using the wrong names affected Windows and not Linux.