UNUSED() compliance for Linux builds, given "teeth" in C++ builds
Though unused variables are removed from the build, their presence can
be a signal of a programming error or oversight.  Or they can just be
plain confusing.

On the path to implementing actual "rigorous" builds across Travis,
this does some of the compliance for turning on the unused variables
warning.  But also, since once an UNUSED() marker goes in, it might
stay on past its lifetime, this employs some C++11 voodoo to corrupt
any local variables that are unused.

(Note: this corrupts for instance the *pointers*, not the things they
are pointing to.)