Turn ROOT and TASK contexts into just ROOT and TASK arrays
R3-Alpha created two notoriously unique "contexts" for the root and
task variables.  These non-user-exposed contexts did not have keylists,
because they were produced before the object-generation machinery
was initialized.  Yet the contexts were not intended to be user
exposed to select values out of.

Much simpler would be just to have them be arrays and not contexts, as
the means of selecting out of them is done by C code.  This does that
change, and also attempts to do some layering so that it's clear what
is "program global" and "task local".  However that is very difficult,
as tasks were never clearly designed in terms of how independent
threads could safely be operating on common memory pools...so this
exposes the dodgy character of how un-separated the task and program
systems actually are.

Also, the values are initialized to unreadable blanks by default, and
only those values that are used by the C get initialized.  This helps
to notice when a root or task variable is no longer used, because the
failure to fill it in will give an assert in ASSERT_ARRAY.