Two Android fixes for RIGOROUS=yes
When creating processes, the `errno` variable was passed by address
to write().  This is not legal on Android, as errno is volatile.
This copies errno to a temporary and writes that instead.

Also, the Android NDK defines char as unsigned by default:

http://stackoverflow.com/questions/7414355/

While most of Rebol uses REBYTE and only uses char for unencoded
string literals, an example in the host was using a < 0 test on char
to tell if it was a UTF8 multibyte character.  This changes that host
code to use REBYTE and check for > 127.