Make C++11 builds test `-funsigned-char`
The C and C++ standards do not dictate if `char` is signed or unsigned.
Lest anyone think all environments have settled on them being signed,
they're not... Android NDK uses unsigned:

http://stackoverflow.com/questions/7414355/

In order to give the option some exercise, make the C++11 builds and
above use unsigned chars.

This expands and corrects the attempt to change to REBYTE in
%host-readline.c from the previous commit.