Simplify HIJACK operation w/zero dispatch overhead, COPY of FUNCTION!
This change radically simplifies HIJACK, in the process making a
hijacked function dispatch with the same performance as the function
which is doing the hijacking.

It also provides a meaningful semantic for COPY of FUNCTION!, which
had historical problems in R3-Alpha.  The new semantic is to do the
proxying which was previously performed with `hijack :some-func _`,
where the idea is that a copy of a function will behave like the
original but with the difference that it has a new identity.  This
new identity would have to be hijacked distinctly from the original.