Rethink previous HIJACK commit in light of problems
Further experimenting with HIJACK revealed that direct modification of
the paramlist was not a good idea, due to the potential dependencies
other locations (specializations, adaptations, MAKE FRAME!s, running
function calls...) might depend on the shape of that paramlist not
changing.

This takes back the idea of doing a replacement of the function body
only if the underlying paramlists are compatible...which will be true
for many adaptations, chains, or specializations from copies.  It is
also true for taking the function back from its original--which solves
the issue of un-hijacking (one of the draws of the approach in the
previous commit, but this way works).

However, a non-compatible underlying function means that the frames
will need to be proxied--effectively making a hijack cost similar to
an extra layer of function call.

At an intermediate point in trying to solve this, it was noticed that
f->func is a somewhat misleading name for the frame's current function,
because it changes.  This changes it to f->phase and preserves the
original function in f->original.

Adds a non-trivial specialization test to hijack tests.