Customizable userspace ECHO based on HIJACK
The ECHO feature was based on some hackish C code inside the device
model.  Given the general desire to simplify and replace the current
workings of that model, this moves the feature into userspace and
bases it on HIJACK of WRITE-STDOUT and INPUT.

As a demonstration of the improved power, this gives ECHO some slightly
interesting features.

The words supported are ON, OFF, and RESET.  You can also specify a
filename or a filename template as a block, where * indicates a place
to substitute in a string.  Supplying a string instead of a filename
then will offer a new substitution.

Sample transcript:

    >> echo on

    >> 1 + 2
    == 3

    >> 3 + 4
    == 7

    >> echo off

    >> read/lines %echo.txt
    == [
        ""
        ">> 1 + 2"
        "== 3"
        ""
        ">> 3 + 4"
        "== 7"
        ""
        ">> echo off"
    ]

    >> echo "cat"

    >> 'meow 'meow
    == meow

    >> echo off

    >> read/lines %echo-cat.txt
    == [
        ""
        ">> 'meow 'meow"
        "== meow"
        ""
        ">> echo off"
    ]

    >> echo [%custom * %-format.log]

    >> echo on

    >> "use blocks for templates"
    == "use blocks for templates"

    >> echo off

    >> read/lines %custom-cat-format.log
    == [
        ""
        ">> echo on"
        ""
        {>> "use blocks for templates"}
        {== "use blocks for templates"}
        ""
        ">> echo off"
    ]

    >> echo reset

    >> read/lines %custom-cat-format.log
    == []

Also puts static into the linker flags, not the compile flags.