Fix HTTPS false alarm error
When porting the original and tough-to-follow HTTPS extension code,
there was a switch() statement that had two cases it checked, and
skipped all else:

https://github.com/zsx/r3/blob/ee00c3dd2295d42e4fec31b389ed841ea21172f0/src/os/host-core.c#L570

In the interest of tightening things up, an error was added to see if
it ever triggered.  It turns out that on certain configurations, it
can trigger because there is a "G" and "PUB-KEY" field in the object.
One instance of causing this was reading https://matrix.org

This relaxes the error to account for those 2 fields, but still gives
an error if unexpected fields are present in the object, which helps
make it more clear what the code is doing.