voiding SELECT*, PICK*, TAKE* w/convenience wraps SELECT, PICK, TAKE
Ren-C introduced the idea that SELECT, in order to be "truthful" about
the full range of values that can appear in a block, would use an
out-of-band signal that cannot appear in a block (void) to denote
that nothing was found. Since BLANK!s can appear in the block, it
wouldn't give you the full information to return blank, since these
two would be indistinguishable:

    >> select [a b c _] 'c
    == _

    >> select [a b] 'c
    == _

Empirical experience of Rebol users in history, however, is that while
"incomplete", the handling of voids is awkward. Frequently it is
known that a block does not contain literal falses or blanks, and
one might casually write:

    >> if x: select [...] y [code]
    Instead of needing to say:

    >> if value? x: select [...] y [code]

(Here, VALUE? is being used as it is intended to someday be retaken,
to mean NOT VOID?)

With other changes to the mechanics of things like if true [] => _
and if* true [] => void, the idea of getting convenient effects
by default suggested that a similar tactic might serve well in
SELECT. Those who know they need the discernment could use SELECT*
and preserve the void/blank distinction, while casual users could
get by with blanks.

This changes the fundamental operations to be the voiding versions,
since they give the most information.  Then it builds convenience
wrappers on top of those, using CHAIN.  The wrapper for TAKE, rather
than converting to a blank when there's nothing to take, gives an
error...because operations that have side effects should generally not
quietly fail.