Add "phase" to FRAME!, fix specializations of CHAINs
Mechanically, the specialization dispatcher was determining what the
function it was specializing was based on the keylist of the exemplar
frame (which is a function paramlist).  However, this keylist is for
the *underlying* function, which was not necessarily the one that
was passed to SPECIALIZE.

This is related to a known problem that since a frame can be used for
many different "phases" of a function, each REBVAL of a frame must
indicate which phase it is for--in order to show the proper subset of
the keylist should be visible to that frame.  By making the phase of
the exemplar frame the function that was specialized, the memory is
retained...then that is used by the dispatcher.