Polymorphic fail() => takes UTF-8 C strings, REBVAL*, REBCTX*
The concept in Rebol is that errors are structural entities, which
hold the relevant parameters for the error, that the user can then
TRAP and extract and react to.  This means that they are defined as
templates in %errors.r, and historically it has required special
syntax from the C code to trigger them.  This also would theoretically
help make it easier to provide internationalization of the strings.

Because the process of adding an error to that table was somewhat of
an effort, especially for an error in transitional code, a "catch-all"
error called RE_MISC was used.  While completely un-informative, it
would serve as a placeholder to come back and replace with a better
error later.

Yet the uninformativeness of an RE_MISC error usually necessitated a
comment anyway.  Much more useful would be to just allow fail() to
accept a C character string and write the comment in that space.  It
can still serve as a sign to come back and fix later with a real
message just as well--just look for `fail ("`.

So this commit eliminates RE_MISC, and takes advantage of the ability
to distinguish UTF-8 strings from REBSER* to allow character strings to
be passed to fail.  As an additional convenience, it takes the #1
most meaningless error about a value (Error_Invalid_Arg) and lets that
lazy option be said even more lazily by passing a REBVAL* to fail():

    fail (Error_Invalid_Arg(ARG(foo)) => fail (ARG(foo))

While perhaps encouraging more laziness, it also makes it easier to
spot that no "real error" is being given in these cases.  A C++
template trick is used to check at compile-time that the void *
parameter isn't actually a RELVAL* or other type Fail_Core() can't
actually handle.