Get rid of Is_Array_Series()
In R3-Alpha, the way that it was determined if a series was an array
or not was if the width was the size of a REBVAL.  That didn't allow
other REBSER which had contents the size of a REBVAL, which might be
useful.  It was changed to a series header bit.  This is now more or
less set in stone, and unlikely to change.

There are a few good reasons to get rid of the abstraction
Is_Array_Series(), which is just a call to GET_SER_FLAG for
SERIES_FLAG_ARRAY.  If people writing routines realize that it is
simply a test of a header flag, then it can make conditional code which
tests other header flags be more obviously finessed by testing multiple
flags at once, or otherwise using the knowledge.

Also, there is a NOT_SER_FLAG variation, which can be more obviously
negated than !Is_Array_Series().

In any case, Is_Array_Series() is not the kind of thing that would
have been added today vs. just testing the series flag.  So that's a
good argument for taking it out.