GC Guard catalogs during boot
The catalog arrays didn't contain anything that had to worry about
GC, but an upcoming commit does have some things that do.  They must
be protected from GC one way or another, so they can be gathered and
then put in the system object later.

Includes misc. comment cleanups, and a fix of an `if` statement that
didn't have parentheses, but got away with it because it used a macro
that would generate the parentheses.