Get rid of "basic LOAD"
There were two versions of LOAD written in Rebol code.  One was the
"early boot version" and the other was the "late boot version".  This
has led to mistakes and misunderstandings, not realizing there are
two LOADs...and as it turns out, the early boot version is never
used.

Removing it for now, and if it needs to be brought back it needs
another name, and an explanation of what it's for.