Reorganize series flags in anticipation of eliminating MKS_XXX
The Make_Series() function takes some flags to dictate whether the
resulting series should be an array, or allocated to a power of 2, etc.

This change moves the SERIES_FLAG_XXX that would be needed to replace
the MKS flags into the same group.  This way, the header could be
assigned directly from the flags, and the series creation behaviors
could then follow from that.

(Rearrangement of the flags is broken out and tested in a separate
commit in order to help isolate problems in the rest of the change.)

Includes a C++11 type checker for the AS_FUNC() cast.