Eliminate MKS_XXX flags in favor of direct SERIES_FLAG_XXX use
This commit streamlines the creation of series slightly, by making the
flags which cue a series creation behavior the exact same flags that
are set in the series header.

Hence, rather than being in a situation like:

    s->header.bits = NODE_FLAG_VALID;
    if (flags & MKS_ARRAY) {
        s->header.bits |= SERIES_FLAG_ARRAY;
        ...
    }

The code can just be like:

     s->header.bits = NODE_FLAG_VALID | flags;
     if (flags & SERIES_FLAG_ARRAY) {
         ...
     }

This also means that many flags that were set after creation can be
passed in as well, saving extra steps.

Default versions of Make_Series() and Make_Array() that do not take
flags are provided as well.

Includes some of the very basic and minimal groundwork for getting file
and line information out of source ANY-ARRAYs.