Fix STRUCT! symbol comparison problem exposed by Haiku build
Because the UTF-8 data behind words is garbage collected, not all
ANY-WORD!s have stable integer symbol values.  Some do, for convenience
and speed in switch() statements inside the C, but others will return
SYM_0 from VAL_WORD_SYM and have to have their Canon form handled
as a varying pointer.

This means that comparing VAL_WORD_SYM() vs VAL_WORD_SYM() and getting
an equal result can be misleading... they could be equal if they are
both REB_0, but not actually the same word.  C++ type tricks are used
to disable equality and inequality comparison directly, requiring
other routines.

The Haiku build noted a bad conversion of an unsigned int to a REBSYM,
and that triggered an error in a strict build which caught such a
comparison in the FFI.