Init_XXX/Shutdown_XXX => Startup_XXX/Shutdown_XXX
There's a lot of terminology overload, and very early it was seen that
if you initialize an integer REBVAL cell with SET_INTEGER(...) then
you get into trouble later when you read things like SET_WORD(...),
because that gets hard to distinguish from `set word value` in the
usermode code.  For this reason, Ren-C started looking toward taking
the word "Init" to be used for initializing value cells.

However, that has a kind of collision with "Init" meaning "Startup",
especially for routines that have a paired "Shutdown".  This makes
things less confusing by taking the functions that don't have to do
solely with initializing value cells, and have a corresponding
Shutdown, and renaming them to a "Startup" prefix.