Q for quitting console compromise, upgrade, non-static default
This implements Q for QUIT as a console-specific instruction, and
also provides an informative message if there is a Q variable that
the user might have thought they wanted to inspect.

By specifically recognizing the console pattern of Q, this avoids
problems like `append [w x y] q` making the console disappear--which
is what would happen with a simpler assignment like `q: :quit`.

Also changes the UPGRADE function definition from "unavailable" to
DO <UPGRADE>, and makes builds non-static by default, which makes the
build work on more platforms by default (clang, but others too).

Moves the ECHO function into the HOST-REPL file as recognition that
a truly good ECHO will need to be interwoven with the console itself,
e.g. to recognize when to take out `>>` prompts in some variants.