Make symbol and boot words interning startup more clear
The terminology used in Startup w.r.t. words and symbols was unclear.
What was actually being initialized by Init_Words() was the string
interning logic, prior to the scanner being able to run...then symbol
initialization was done afterwards.  This organizes and renames the
code to make that more clear, un-comments asserts to either enable
them or to put them under #ifdefs.