Rewrite and frame optimization of varargs, DECLARE_FRAME macro
An assert added to Do_Next_In_Subframe_Throws() detected situations
where the creation of a subframe was wasteful.  The VARARGS! code, when
operating on an array, was known to be a little wasteful by creating
frames even in quoting scenarios...just to present a compatible
interface to the frame-based VARARGS! and not have to write two
versions of that code.  The assert pointed out this was doubly wasteful,
because this frame held no state that required a subframe in order to
not overwrite.

This indicated that the common code was less common than thought, and
hence a new form of factorization was introduced that allowed the two
kinds of VARARGS! to share as much as possible while having different
forms of frame access.  This motivated a larger-scale cleanup, which
changed the interface of Do_Vararg_Op_May_Throw to one that uses
R_XXX values instead of REBIXO flags, and a subroutine that makes use
of the R_UNHANDLED value in the enumeration.

Desire to have better handling of the frame's internal cell preparation
and to make code that declares stack frames speak in terms of pointers
motivated the creation of a DECLARE_FRAME macro, inspired by
DECLARE_LOCAL.  This meant turning a lot of `f.xxx` accesses into
`f->xxx` accesses.

Also fixes the behavior of |> and <|, which were being implemented
with TAKE* and using a void signal to detect the end, which is not
applicable to them because they tolerate voids...so TAIL? must be
used to discern a true end from a void evaluation.