Disallow PICK on MAP!
This should never have been allowed, but R3-Alpha allowed it.  What
it does is expose the guts of the MAP! implementation, which has no
bearing on what the user can meaningfully process.

Most likely the reason it was there is because Rebol2 supported an
accelerated index over BLOCK! known as HASH!, and it retained the
semantics of a BLOCK!.