Allow hard-quoting of BAR!, yet still disallow QUOTE |
The desire to give BAR! more "teeth" led to the idea that it could
not be quoted, *not even hard-quoted*.  This was based on the idea that
code which said something like `[quote | quote]` should not have the
first quote see any less of an ending to the stream of arguments than
the second quote.

Several use cases, in particular applicable to variadics, show that
it can be useful to see a | as distinct from the end of a block.  The
constraint of only allowing this for hard quotes seems restrictive
enough, especially given the fact that any given hard quoting operator
can within its own implementation reject a literal BAR!.  This is
exactly the behavior grafted into QUOTE.

As a practical demonstration, |> and <| are modified to see past BAR!

   >> 1 + 2 <| 30 + 40 | print "<| is still running" | print "still"
   <| is still running
   still
   3