Tidy up %prot-tls.r for readability / state isolation
This is some idle cleanup done while trying to read through and
understand what the TLS code was doing, in order to trace a bug.

It seemed that it was being "stateful" somehow and having one
transaction corrupt another, so this went ahead and turned the FUNCs
into FUNCTION.  Moved some globals that were used only by one routine
into static variables, and moveto being a little more compliant with
the 80 character width limit.