Fix access of END markers in copy for negative indexes
The code `copy/part tail [1] -1` appeared to work in the tests, but
was actually reading memory it shouldn't...because it wasn't taking
into account the change of index by Partial1.  So in that example, it
was actually trying to byte-copy two values, one of which was an
internal end marker for a singular array.