Change head of object keylists to unreadable blank vs. trash
Previously, the keylist building routine would put a "trash" cell in
the [0] slot, later overwritten with a plain blank to appease the
garbage collector.  Since this slot is reserved for future use, an
unreadable blank is better...but is also should not spend time being
a trash cell, because that means the copy routines would have to
tolerate trash.

Includes removal of a superfluous binding parameter to a context
creation routine, and using Derive_Specifier in path dispatch instead
of repeating that code.