Rearrange datatypes for "bindable" split
With binding becoming a fundamental property that datatypes either
do or do not have, it is checked on movement and derelativization.
This means it happens frequently, so one does not want to do a large
set of `if` statements for many different classes.  Better than setting
a bit is to actually be able to determine it from the type value itself
being greater than some value (since a bit would have to be set through
a table or some other means, while this is intrinsic to the number.)

There is limited meaning that can be encoded in the actual numeric
values chosen for datatypes.  What that might be used for has been
experimented with over time.  (One of the bigger outcomes of that was
making it safer to experimented with and change more often.)

Since R3-Alpha, classes have been grouped together...e.g. all the
ANY-SERIES types are sequential, all the ANY-ARRAY!, etc.  An early
attempt at optimization arranged the types so that REB_0 and
REB_FUNCTION were at the head so they could be interchangeable with
boolean logic, but that optimization was made obsolete by
VALUE_FLAG_ENFIXED and Move_Value().  Another attempt was to arrange
all the evaluative types before all the non-evaluative types in order
to determine when to make a recursion, but the nuances of infix
processing have made that optimization extremely difficult to do in
a way that adds value.

Given the frequency of the binding check, it needs to be fast, and
the numbering is currently adjustable.  This pushes the numbers around
so all the bindable values are below a certain number.  (They're all
less than REB_BAR.)  Some types are bindable but don't need to be,
e.g. REB_IMAGE and REB_VECTOR, but it's not worth changing their
implementations to not be ANY-SERIES! at this moment in time.