Make #ifdef unnecessary by splitting into two lines, comment updates
The difference between C and C++ w.r.t. const declarations is such
that `extern` can't be applied to a declaration at the point of its
initialization.  Rather than use an #ifdef, this just splits into
a forward declaration and then the initialized declaration.

Includes some thinning out of repetitive comments in %sys-value.h
regarding the track payload, while making a greater case for its usage
by emphasizing it at the top of the file.