Beginnings of userspace encap, ELF format add/update, zip
This is the start of work on improving encapping, by removing it from
the C code and putting it into the host, such that the encap tool and
the encap checks are done by platform-independent Rebol code.  Hence
a Windows Rebol executable could encap a Linux Rebol executable that
was acquired from somewhere else.

Though this commit only contains code for ELF, @ShixinZeng is working
on a Windows PE version.  It contains a fallback for a "naive"
appending mode for executable formats it does not understand.

This includes a patched version of the %unzip.r from rebol.org to work
under Ren-C into the host, and a simple first cut at interoperating
with it in order to unpack a compressed directory into the
system/options/encap as a block.  The assumption is that the script
to run is called %main.reb.  A better level of abstraction, ideally
through URL! like `encap://subdir/resource.png` would be desired,
with the ability to selectively specifiy whether resources should be
loaded from disk or from inside the executable.  Evolutions of this
strategy could be used to run the mezzanine from disk, if it too
were included as part of the encap.