Resolve problem with DS_TOP asserts, bolster DS_AT assert
The DS_TOP inline function was modified to ensure that the top of stack
was not an END marker.  However, END markers are the signal for
when an expansion should happen.  The "built-in" test of stack
expansion was modified to account for the assert, while the plain
stack expansion was not.

This removes usages of DS_TOP and DS_AT from the internal stack
expansion logic, where END markers are possible...and strengthens
DS_AT by making it not be willing to return END markers either.

Also corrects an issue with the STRIP tool aliasing noticed on MinGW.