Fix invalid use of VAL_TYPE_RAW() on what could be an internal END
In order to speed up some usages of VAL_TYPE, they were substituted
with VAL_TYPE_RAW, with END cells and the global END marker giving
back REB_0 as the type.

Yet an oversight had an instance of using this trick on a REBVAL* that
was not known to be a cell end, but could be an internal "endlike"
header.  The problem is that the area corresponding to the type bits
for such pointers could be anything.  Hence the trick cannot be used
if you are doing a traversal and do not know it's not a cell or the
global END.

Because it's rare to know this specifically, this moves to a safer
choice than VAL_TYPE_RAW(), with VAL_TYPE_OR_0().  This has all the
checking of VAL_TYPE(), except makes a special exception for the
global end, returning REB_0.

This also moves around some routines to be inline and close to the
definitions they are related to, due to some dependencies now no
longer being missing in order to write those constructs in headers.
It also makes trash cells use REB_MAX + 1 for the type in the debug
build.