LENGTH-OF, HEAD-OF, TAIL-OF as core names...aliased to shorter
Controversy regarding the "verbing" of nouns has led to questions of
what the appropriate abbreviation of the length, head, and tail
operations are.

This commit hedges, by making Rebol internally use long names, but
then aliasing them to short names.  This means if someone happens to
use something like HEAD or TAIL as a variable, they can get back to
the operation via using HEAD-OF or TAIL-OF.  Same for LENGTH.