Make APPEND on port act as WRITE/APPEND in slightly less hacky way
R3-Alpha implemented the APPEND action inside of a FILE!-based port
specifically, in order to make it possible to APPEND to something
that might be a BINARY! or a STRING!.  In doing so, it assumed that
WRITE and APPEND had the same frame--which is not necessarily the
case, and is an unnatural assumption.  It was not working with the
new INCLUDE_PARAMS_OF_XXX macros which verify the frame is compatible
with what is expected.

This decouples the frame expectation by actually re-triggering an
APPEND on a FILE!-based PORT! through WRITE/APPEND.  This is a bit
hacky, but a workaround that keeps code that depended on this going
for the moment.  A general review of the semantics of APPEND on a
PORT! needs to happen, because since a PORT! is an ANY-CONTEXT! this
fundamentally overrides the behavior of APPEND as expected on objects,
and is a sort of basic question about how a port is expected to act
as a context for its internal implementation as well as act as a
sort of arbitrary black box for clients using the same kinds of
operations.