assert=>error on `eval ()`, patch EVAL/ONLY for SET-WORD/PATH!
The notion of "evaluating a void" is not something that comes up
naturally when dealing with a block, since there are no voids in blocks
as cells.  However, EVAL is able to produce the situation, and it can
also come up if a bad client using the va_list interface passes in a
void as evaluative input.  So this notices those cases and gives an
appropriate error, while tolerating the non-evaluative case.

Also patches EVAL/ONLY which had stopped working for SET-WORD or
SET-PATH!, e.g.:

    >> eval/only quote x: 1 + 2
    == 3

    >> x
    1